Adding support for loading static resources (image, css, javascript) from the plugin file. No extraction is needed; instead, a generic way of loading resource inside jar file is introduced by utilising Weblets (https://weblets.dev.java.net/). In general, if a plugin needs to use self-supplied static resources (i.e., resources available inside the plugin jar file itself), it will have to declare itself also as a weblet by providing a file named weblets-config.xml located inside META-INF folder. An example of such config file is as following
<?xml version="1.0" encoding="UTF-8" ?>
<weblets-config xmlns="http://weblets.dev.java.net/config">
  <weblet>
    <weblet-name>org.example.hello</weblet-name>
    <weblet-class>net.java.dev.weblets.packaged.PackagedWeblet</weblet-class>
    <init-param>
      <param-name>package</param-name>
      <param-value>resources</param-value>
    </init-param>
  </weblet>
  <weblet-mapping>
    <weblet-name>org.example.hello</weblet-name>
    <url-pattern>/example_hello/*</url-pattern>
  </weblet-mapping>
</weblets-config>

    o The <weblet-name>-element provides a unique name for this weblet. It's
      therefore recommended to use the pluginId as weblet name.
    o The <weblet-class>-element should be exactly as given since we are
      using what called PackagedWeblet (resource packaged inside jar
      file). This suggests there're other uses of weblet; however, we
      will not be concerned with these for the purpose of our plugin
      system.
    o The <init-param>-element is required for packaged weblet. The
      only value the plugin developer has to pay attention to is
      <param-value> of the package parameter. This specify the
      package inside the plugin jar file where we can find the
      static resources. In this example we put all our static resources
      file in a package named resources (i.e., resources folder at root
      of the jar file).
    o The <weblet-mapping> defines the mapping of urls to corresponding
      resources.

E.g.
Suppose we have this weblet config in a plugin jar file and we have an
image located as

        /resources/img/sample.jpg

Under this configuration, we can load this image by this URL

        http://<webapp-url>/weblets/weblets/example_hello/img/sample.jpg

Notice how this setting reflects the configuration we have in the config
file. However, as spelling out the whole url is a real pain in some
cases, we have a helper method to get this exact link. When writing a
Velocity template file, one always has access to the PageContext
instance under variable $context. A method named
getWebletResourceURL(String weblet, String pathInfo) is available on the
PageContext to provide links to resouces inside a weblet. For example,
if we want to get link to the image above, we only need to call

    $context.getWebletResourceURL("org.example.hello", "/img/sample.jpg")

The first argument is obvioulsy the name of the weblet which we have
defined and the second argument is the path and filename of the resource
which we want to load.

Also note that this getWebletResourceURL method is simply a wrapper call
to WebletUtils.getURL method provided by the Weblet implementation.
Furthermore, to introduce Weblet to the current system, apart from some
new lib and dependencies, we also have to define a new designated Weblet
servlet. Take a look at the diff of web.xml to see how this new servlet
is defined.

For more information on Weblet, feel free to visit the project's website.