Refactoring the extension point for plugins to provide new Database Access Object (DAO). Previously, a plugin has to extend MySQLPluginEntityDAO (which is a subclass of MySQLEntityDAO). While this approach is very natural given that all the entities in the core BOSS system were created this way, it is inherently unsafe to let external code extends this class. Doing so will result in external class having access to the current working java.sql.Connection object (since this is a member field of every MySQLEntityDAO) and therefore will be able to get the database connection details and possibly execute arbitrary SQL statements on the core system's database.
This commit refactors the code so that plugin DAO object will have to
extend PluginEntityDAO which is essentially a simple interface
specifying the contracted methods needed for a basic entity DAO object.
This approach is safer since PluginEntityDAO is a rather independent
class which has no knowledge of the database connection details.
Implementations of these contracted methods will be in turned called
from MySQLPluginEntityDAO which is essentially a wrapper around
PluginEntityDAO. This wrapper is indeed a proper entity DAO class (i.e.,
subclass of MySQLEntityDAO) and therefore will be able to function as
one. However, this class is not an extension point therefore plugin code
can't really inherit or access the connection details resided inside
this class.