Modify the code to utilise the new approach of creating and deleting plugin custom table. With this new commit, when a plugin create a new entity and provide a new entity DAO for it, it will have to implement an additional method called getTableSchema which will return a SQLTableSchema object representing the schema of the underlying table of this new entity. This method is now part of the PluginEntityDAO SPI.
The installation method will try to load the appropriate implementation
of PluginEntityDAO and call getTableSchema on this to obtain the schema
object. This can then be used to get the corresponding CREATE TABLE
string and execute. The problem however is at installation, the
implementation of PluginEntityDAO in this plugin is not in the CLASSPATH
yet (the CLASSPATH will be updated after the installation is finished
and the admin has restarted the application). So, the default lookup
will not be able to load this implementation. The solution for this was
we create a temporary custom ClassLoader which contains the plugin Jar
file. We then use lookup with this custom ClassLoader. Now, the
implementation can be loaded but we have another problem in which if we
have previously installed plugin who also provide implementation of
PluginEntityDAO, we'll not be able to tell which is which and might end
up execute the creation script of some other plugin. The solution for
this one is quite subtle. We will base on different ClassLoader objects
of the implementations to distinguish them. Let say the implementation
which is currently loaded is pluginDb, we can get the ClassLoader object
of pluginDb by calling pluginDb.getClass().getClassLoader(). Under the
delegation model of Java's ClassLoader, if pluginDb is the one belongs
to the previously installed plugin, its ClassLoader would be the current
thread context class loader while if it is the one we are looking for
(the one loaded from the plugin Jar file), its classloader would be the
exact same one we have created earlier. So, a simple comparison will be
able to tell if we got the right implementation.

On the other hand, in uninstallation process, we would want to do the
same thing, if we could load the correct PluginEntityDAO implementation
of the plugin which we want to uninstall, we can call getTableName on
that to get the name of the custom plugin table. The we can simply form
a DROP TABLE <tableName> statement and execute. The problem we faced
here is similar to the installation process, we can't easily tell which
implementations of PluginEntityDAO belong to the current uninstalling
plugin using Lookup. However, in here the situation is a little bit
different comparing to the installation case because right now
essentially all implementations of PluginEntityDAO are loaded from the
same ClassLoader (i.e., the current thread context class loader) so the
trick described previously will not work. Fortunately, there's a way we
could find out where exactly in the file system that this implementation
has been loaded. Assume will have the current implementation in question
as pluginDb, the following call will give the full path to the jar file
where this implementation is loaded

pluginDb.getProtectionDomain().getCodeSource().getLocation().getPath()

Now since we are looking for the ones belong to a plugin with a unique
pluginId and we know that our plugin system stored this plugin's jar
file under <WebAppContextPath>/WEB-INF/lib/plugin_<pluginId>.jar, we can
easily tell which plugin that pluginDb belongs to. We can then proceed
to drop the corresponding tables.